pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["EfremovAS <Alexef3@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.114.0"}
uvicorn = "^0.30.6"
pydantic = "^2.9.1"
pydantic-settings = "^2.5.2"
sqlalchemy = "^2.0.34"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
black = "^24.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.9.0"
redis = "^5.0.8"
fastapi-cache2 = "^0.2.2"
celery = "^5.4.0"
pillow = "^10.4.0"
pytest = "^8.3.3"
pytest-dotenv = "^0.5.2"
pytest-asyncio = "^0.24.0"
ruff = "^0.6.8"
pyright = "^1.1.382.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
